style(CalculatorController): Application.java에 있던 컨트롤러를 컨트롤러 패키지 만들어서  분리
변경 전
- Application 내부에 CalculatorController 존재

변경 후
- calculator 패키지에 controller 패키지를 만들고, 그 안에 CalculatorController.java 클래스를 만들어 컨트롤러가 동작할 수 있도록 함
