+++ src/main/java/calculator/Application.java
@@ -1,30 +1,8 @@
 package calculator;
 
+import calculator.controller.CalculatorController;
 import calculator.model.Calculator;
-import calculator.model.Expression;
 import calculator.view.CalculateView;
-
-// controller
-class CalculatorController {
-    private final Calculator model;
-    private final CalculateView view;
-
-    public CalculatorController(Calculator model, CalculateView view) {
-        this.model = model;
-        this.view = view;
-    }
-
-    public void run() {
-        String expression = view.getUserInput();
-
-        // 데이터 처리
-        Expression expressionModel = new Expression(expression);
-        int result = model.calculate(expressionModel);
-
-        // 출력
-        view.displayResult(result);
-    }
-}
 
 public class Application {
     public static void main(String[] args) {
